Bug fixes - clear old data in comms buffers.
Modbus::get_FC1() - the code used to preserve old data in the
high byte of au16regs words. So, if a single coil is returned
from the slave, au16regs[0]'s high byte was left with junk in it.

Modbus::process_FC1() - make sure that the data returned to the
master is clean. Unused bits should be zeroed, according to the
spec.